+++ import.go
@@ -157,10 +157,12 @@ func (i *Importer) Commit() error {
 		return ErrNoImport
 	}
 
-	switch {
-	case len(i.stack) == 1:
+	switch len(i.stack) {
+	case 0:
+		i.batch.Set(i.tree.ndb.rootKey(i.version), []byte{})
+	case 1:
 		i.batch.Set(i.tree.ndb.rootKey(i.version), i.stack[0].hash)
-	case len(i.stack) > 2:
+	default:
 		return errors.Errorf("invalid node structure, found stack size %v when committing",
 			len(i.stack))
 	}