+++ CHANGELOG.md
@@ -10,6 +10,10 @@ Special thanks to external contributors on this release:
 
 ### Bug Fixes
 
+## 0.13.2 (April 5, 2020)
+
+### Bug Fixes
+
 - [import] [\#230](https://github.com/tendermint/iavl/pull/230) Set correct version when committing an empty import.
 
 ## 0.13.2 (March 18, 2020)